+++ src/CrossReferences.jl
@@ -13,6 +13,7 @@ import ..Documenter:
     Utilities.@docerror
 
 using DocStringExtensions
+using .Utilities: Remotes
 import Markdown
 
 """
@@ -213,8 +214,13 @@ getsig(λ::Union{Function, DataType}, typesig) = Base.tuple_type_tail(which(λ, 
 # -----------------------------
 
 function issue_xref(link::Markdown.Link, num, meta, page, doc)
-    link.url = isempty(doc.internal.remote) ? link.url :
-        "https://github.com/$(doc.internal.remote)/issues/$num"
+    # Update issue links starting with a hash, but only if our Remote supports it
+    issue_url = Remotes.issueurl(doc.user.remote, num)
+    if isnothing(issue_url)
+        @docerror(doc, :cross_references, "unable to generate issue reference for '[`#$num`](@ref)' in $(Utilities.locrepr(page.source)).")
+    else
+        link.url = issue_url
+    end
 end
 
 end